add mentor item component
