+++ components/MentorList.tsx
@@ -2,6 +2,7 @@ import { fetchMentors } from "@/lib/firebaseUtils"
 import { FirestoreMentor, FirestoreUser } from "@/lib/types"
 import { useEffect, useState } from "react"
 import Separator from "./Separator"
+import MentorItemComponent from "./MentorItemComponent"
 
 export default function MentorListComponent() {
   const [mentors, setMentors] = useState<FirestoreMentor[]>()
@@ -54,36 +55,7 @@ export default function MentorListComponent() {
           <h1 className="text-xl font-semibold">{s.toUpperCase()}</h1>
           <div className="rounded-xl p-4 grid grid-cols-1 md:grid-cols-2 gap-4">
             {mentors?.filter((ms) => ms.specialization.includes(s)).map((m) => (
-              <div key={m.id} className="border rounded-xl border-gray-400 p-4 flex justify-between">
-                <div className="flex flex-col gap-2">
-                  <p className="font-semibold">{m.name}</p>
-                  <p className="text-muted-foreground">{m.email}</p>
-                  <p className="">{m.specialization.toUpperCase()}</p>
-                  <div className="">
-                    {m.availableMentorings.map((availableMentoring, index) => {
-                      const startDate = new Date(availableMentoring.startTime * 1000)
-                      const startDay = startDate.toLocaleDateString()
-                      const startTimestamp = startDate.toLocaleString('id-ID', { timeStyle: 'short' })
-                      const start = `${startDay} ${startTimestamp}`
-
-                      const endDate = new Date(availableMentoring.startTime * 1000)
-                      const endDay = endDate.toLocaleDateString()
-                      const endTimestamp = endDate.toLocaleString('id-ID', { timeStyle: 'short' })
-                      const end = `${endDay} ${endTimestamp}`
-                      
-                      return (
-                        <div key={index} className="text-sm flex gap-2 px-2 border bg-muted rounded-xl">
-                          <span>{start} WIB - {end} WIB</span>
-                        </div>
-                      )
-                    }
-                    )}
-                  </div>
-                </div>
-                <button className="font-semibold text-sm hover:underline">
-                  View
-                </button>
-              </div>
+              <MentorItemComponent key={m.id} mentor={m} />
             ))}
           </div>
         </div>
